backend/rag_api_service/requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
python-multipart>=0.0.6
boto3>=1.26.0
python-dotenv>=1.0.0